Added DMOJ solution to Dynamic Range Minimum Test
The following code is the solution to Dynamic Range Minimum Test, although it can also function as a relatively easy to understand code for Segment Tree. In this case, the range query implemented is simple, just range minimum query. As well, range update is not implemented in this code(a future file will have that implementation).